Extract tests module to file in vfs crate
